refactor(desktop): extract agent activity opening actions
