+++ desktop/src/features/channels/ui/ChannelScreen.tsx
@@ -1,5 +1,4 @@
 import * as React from "react";
-import { toast } from "sonner";
 import { useAppShell } from "@/app/AppShellContext";
 import { useAppNavigation } from "@/app/navigation/useAppNavigation";
 import { useActiveChannelHeader } from "@/features/channels/useActiveChannelHeader";
@@ -58,7 +57,6 @@ import type { TimelineMessage } from "@/features/messages/types";
 import { useUsersBatchQuery } from "@/features/profile/hooks";
 import { useRelaySelfQuery } from "@/features/moderation/hooks";
 import type { RelayEvent, RespondToMode } from "@/shared/api/types";
-import { openAgentActivityWindow } from "@/shared/api/agentActivityWindow";
 import { isAgentActivityWindow } from "@/features/agents/lib/agentActivityWindow";
 import { ChannelScreenLoadingFallback } from "@/features/channels/ui/ChannelScreenLoadingFallback";
 import {
@@ -79,6 +77,7 @@ import { normalizePubkey } from "@/shared/lib/pubkey";
 import { useChannelActivityTyping } from "./useChannelActivityTyping";
 import { useChannelAgentSessions } from "./useChannelAgentSessions";
 import { useMessageProfiles } from "./useMessageProfiles";
+import { useOpenAgentActivityActions } from "./useOpenAgentActivityActions";
 import { useChannelPanelHistoryState } from "./useChannelPanelHistoryState";
 import { useChannelProfilePanel } from "./useChannelProfilePanel";
 import { useChannelRouteTarget } from "./useChannelRouteTarget";
@@ -587,35 +586,10 @@ export function ChannelScreen({
     setThreadReplyTargetId,
     setThreadScrollTargetId,
   });
-  const handleOpenAgentActivityExternal = React.useCallback(
-    (pubkey: string, channelId?: string | null) => {
-      const destinationChannelId = channelId ?? activeChannelId;
-      if (!destinationChannelId) return;
-
-      if (isAgentActivityWindow()) {
-        handleOpenAgentSession(pubkey, destinationChannelId);
-        return;
-      }
-
-      void openAgentActivityWindow(destinationChannelId, pubkey)
-        .then((openedNativeWindow) => {
-          if (!openedNativeWindow) {
-            handleOpenAgentSession(pubkey, destinationChannelId);
-          }
-        })
-        .catch((error) => {
-          console.error("Failed to open agent activity window:", error);
-          toast.error("Couldn't open the agent activity window.");
-        });
-    },
-    [activeChannelId, handleOpenAgentSession],
-  );
-  const handleOpenAgentActivity = React.useCallback(
-    (pubkey: string, channelId?: string | null) => {
-      handleOpenAgentSession(pubkey, channelId ?? activeChannelId);
-    },
-    [activeChannelId, handleOpenAgentSession],
-  );
+  const {
+    openInApp: handleOpenAgentActivity,
+    openInExternalWindow: handleOpenAgentActivityExternal,
+  } = useOpenAgentActivityActions(activeChannelId, handleOpenAgentSession);
   const { handleOpenProfilePanel, handleCloseProfilePanel, handleOpenDm } =
     useChannelProfilePanel({
       closeAgentSession: handleCloseAgentSession,
